Refactor mypyc singledispatch support
Move all of the singledispatch-related functions in
mypyc/irbuild/function.py into a separate file just for
singledispatch-related functionality.

This also adds documentation about how mypyc's support for
singledispatch works.
